requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.9
frozenlist==1.2.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
line-bot-sdk==2.0.1
multidict==5.2.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
django-import-export==2.6.1
